requirements.txt
numpy==1.21.1
opencv-python==4.5.3.56
pkg-resources==0.0.0
pygame==2.0.1
RPi.GPIO==0.7.0
